检视意见修改，去掉多余函数

+++ rosen/modules/render_service_client/core/ui/rs_node.cpp
@@ -868,7 +868,7 @@ void RSNode::SetBgImagePositionY(float positionY)
 void RSNode::SetColorBlendMode(const RSColorBlendModeType blendMode)
 {
     SetProperty<RSColorBlendModeModifier, RSProperty<RSColorBlendModeType>>(
-        RSModifierType::COLOR_BLENDMODE, blendMode);
+        RSModifierType::COLOR_BLENDMODE, static_cast<int>(blendMode);
 }
 
 // border
